+++ .gitignore
@@ -4,4 +4,5 @@ dist
 test
 package-metrics/node_modules
 package-metrics/build
+./logs
 coverage
